LCM: work on SchwarzLoopDynamics.
Pushing relatively small changes to avoid conflicts with AM's commits in the future.

Adding in call to writeSolution for solution observation with Tempus, though this design
may need to change as I am not sure that all the states / stresses will be written to Exodus
with it.

Adding some debug output.