LCM: temporary solution to write to Exodus file for SchwarzLoopDynamics.
For some reason, with Tempus, the solution does not get written to the stk
discretization object, whereas it does for LOCA.  I suspect this may have
to do with the observer.  With this fix, solution and its derivatives
et written to Exodus but not the Cauchy stresses, so it is temporary to help
with debugging for now.