LCM: changing how prev_soln_rcp and its derivatives is set in SchwarzLoopDynamics.
In particular, making it a copy of solutions_thyra_[subdomain]
rather than setting the pointers equal to each other. This is necessary
because the object passed to setInitialState gets modified inside
the time-integration.

There is still a problem with SchwarzLoopDynamics, as code does not
converge for simple cuboid problem with steady BCs.

Adding debug output in the form of dumping vectors to matrix market.